backend/requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.10
django-cors-headers==4.2.0
django-postgres-extensions==0.9.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
packaging==23.1
Pillow==10.0.0
PyJWT==1.7.1
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.9.2
sqlparse==0.4.4
swagger-spec-validator==3.0.3
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.4
